Code correction

+++ Coditech.Project/Common/Coditech.Common.API.Model/GeneralMaster/GeneralNationalityMaster/GeneralNationalityModel.cs
@@ -8,7 +8,7 @@ namespace Coditech.Common.API.Model
         {
 
         }
-        public int GeneralNationalityMasterId { get; set; }
+        public short GeneralNationalityMasterId { get; set; }
         public string Description { get; set; }
         public bool DefaultFlag { get; set; }
     }
